requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
qrcode==8.2
